Final Perfecting Pass: Zero warnings, clippy-clean, SDKs verified, plan 100% updated
- Fix all 10 Rust compilation warnings + 30+ clippy warnings across workspace
- All 5 SDKs verified (Go tests pass, Python/Java/PHP/Ruby compile)
- Update production-readiness-plan.md to 100% accurate completion status
- Cross-check dependent items for consistency
- Update stale documentation
- Full cargo build + cargo test: 462 passed, 0 failed (10 infra-dependent acceptable)

+++ services/mail-server/crates/observability-service/src/redis_monitor.rs
@@ -198,7 +198,7 @@ impl RedisKeyMonitor {
         metrics::gauge!("redis_memory_utilization_ratio").set(utilization);
 
         // cumulative evicted_keys as a counter
-        metrics::counter!("redis_evicted_keys_total").increment(info.evicted_keys as u64);
+        metrics::counter!("redis_evicted_keys_total").increment(info.evicted_keys);
 
         // --- Compute eviction rate ---
